LASB-798: updated joson schema, pojo files as well as rearranged Means assessment controller related funcitonality and its test
