+++ crime-means-assessment/src/main/java/uk/gov/justice/laa/crime/meansassessment/model/common/ApiAssessmentSummary.java
@@ -20,11 +20,11 @@ import java.util.List;
  * Assessment Summary
  * 
  */
-@Generated("jsonschema2pojo")
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
+@Generated("jsonschema2pojo")
 public class ApiAssessmentSummary {
 
     /**
