+++ crime-means-assessment/src/main/java/uk/gov/justice/laa/crime/meansassessment/model/common/ApiReviewType.java
@@ -3,6 +3,10 @@ package uk.gov.justice.laa.crime.meansassessment.model.common;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.annotation.processing.Generated;
 import javax.validation.constraints.NotNull;
@@ -12,6 +16,10 @@ import javax.validation.constraints.NotNull;
  * An explanation about the purpose of this instance.
  * 
  */
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
 @Generated("jsonschema2pojo")
 public class ApiReviewType {
