+++ crime-means-assessment/src/test/java/uk/gov/justice/laa/crime/meansassessment/data/builder/TestModelDataBuilder.java
@@ -146,6 +146,8 @@ public class TestModelDataBuilder {
                 .newWorkReason(getApiNewWorkReason())
                 .supplierInfo(getApiSupplierInfo())
                 .assessmentSummary(getAssessmentSummaries())
+                .hasPartner(true)
+                .partnerContraryInterest(false)
                 .build();
         return meansAssessmentRequest;
     }
